+++ axum-extra/CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning].
 
 # Unreleased
 
+- None.
+
+# 0.3.4 (08. June, 2022)
+
 - **fixed:** Use `impl IntoResponse` less in docs ([#1049])
+- **added:** Add `AsyncReadBody` for creating a body from a `tokio::io::AsyncRead` ([#1072])
 
 [#1049]: https://github.com/tokio-rs/axum/pull/1049
+[#1072]: https://github.com/tokio-rs/axum/pull/1072
 
 # 0.3.3 (18. May, 2022)